update(driverkit): fixed specific_target support for multiarch.
Note: even when filtering for TARGET_ARCH, `validate_old` and `specific_target_old` will always run.
This is a small issue while we support both the old and the new format,
and will eventually go away once we drop support for old format.
It is not impactful since we do never filter for TARGET_ARCH (given that it is a newly introduced filter).

Signed-off-by: Federico Di Pierro <nierro92@gmail.com>